requirements.txt
arch: android-tools
customtkinter
pillow